+++ SceneWithSkull/DemoMain.cpp
@@ -136,7 +136,7 @@ void DemoMain::Update()
     XMVECTOR eye = m_input->GetPosition();
     auto viewMatrix = m_camera->GetViewMatrix(eye);
 
-    m_renderer->Update(viewMatrix, eye, m_timer.GetElapsedSeconds());
+    m_renderer->Update(viewMatrix, eye, static_cast<float>(m_timer.GetElapsedSeconds()));
 }
 
 /// <summary>
